Move API Ref MDX components to mapping
