+++ src/components/Documentation/SharedStyles.js
@@ -1,5 +1,3 @@
-import React from "react";
-import PropTypes from "prop-types";
 import styled from "styled-components";
 
 import { MEDIA_QUERIES } from "constants/styles";
@@ -34,52 +32,6 @@ export const CustomColumn = styled(Column)`
     grid-column: ${(props) => (props.xlColumn ? props.xlColumn : "inherit")};
   }
 `;
-const RIGHT_COLUMN_COMPONENTS_NAME = {
-  CodeExample: "CodeExample",
-  EndpointsTable: "EndpointsTable",
-  ExampleResponse: "ExampleResponse",
-  NavTable: "NavTable",
-};
-// API reference has 2 columns, with certain types of content always being in one
-// or the other. This component sorts through its children to separate them.
-export const ApiReferenceWrapper = ({ children, ...props }) => {
-  const rightColumnContent = React.useMemo(
-    () =>
-      React.Children.toArray(children).filter(
-        (child) => RIGHT_COLUMN_COMPONENTS_NAME[child.props.mdxType],
-      ),
-    [children],
-  );
-  const MiddleColumnContent = React.useMemo(
-    () =>
-      React.Children.toArray(children).filter(
-        (child) => !RIGHT_COLUMN_COMPONENTS_NAME[child.props.mdxType],
-      ),
-    [children],
-  );
-
-  return (
-    <React.Fragment>
-      {/* Hack to make it look appear as if we had a column-gap
-      4rem in between <CustomColumn/> on a large screen (min-width: 1440px)
-      skip the 1st column to use it as column-gap, start at the 2nd column and
-      span through then next 8 columns (ends at column 9) */}
-      <CustomColumn xs={5} xl={9} xlColumn="2 / span 8">
-        <Content {...props}>{MiddleColumnContent}</Content>
-      </CustomColumn>
-      {/* Hack to make it look appear as if we had a column-gap
-      4rem in between <CustomColumn/> on a large screen (min-width: 1440px)
-      skip the 10th column to use it as column-gap, start at the 11th column and
-      span through the next 8 columns (ends at column 18) */}
-      <CustomColumn xs={4} xl={9} xlColumn="11 / span 8">
-        {rightColumnContent}
-      </CustomColumn>
-    </React.Fragment>
-  );
-};
-ApiReferenceWrapper.propTypes = {
-  children: PropTypes.node.isRequired,
-};
 
 const { getSizeGrid, COL_SIZES, COLUMNS } = gridHelpers;
 const { count, size, margin } = COLUMNS[COL_SIZES.md];
